refactor: move integration helpers in integration folder
