+++ tests/integration/conftest.py
@@ -3,9 +3,9 @@ import logging
 import pytest
 from starkware.starknet.testing.contract import StarknetContract
 
+from tests.integration.helpers.helpers import hex_string_to_bytes_array
+from tests.integration.helpers.wrap_kakarot import get_contract, wrap_for_kakarot
 from tests.utils.reporting import traceit
-from tests.utils.utils import hex_string_to_bytes_array
-from tests.utils.wrap_kakarot import get_contract, wrap_for_kakarot
 
 logger = logging.getLogger()
