+++ tests/unit/src/kakarot/accounts/registry/test_account_registry.py
