+++ tests/unit/src/kakarot/instructions/test_push_operations.cairo
@@ -13,7 +13,7 @@ from kakarot.model import model
 from kakarot.stack import Stack
 from kakarot.execution_context import ExecutionContext
 from kakarot.instructions.push_operations import PushOperations
-from tests.utils.utils import TestHelpers
+from tests.unit.helpers.helpers import TestHelpers
 
 @external
 func test__exec_push1_should_add_1_byte_to_stack{
