+++ tests/unit/src/kakarot/test_execution_context.py
@@ -8,7 +8,7 @@ from starkware.starknet.testing.starknet import Starknet
 @pytest_asyncio.fixture(scope="module")
 async def execution_context(starknet: Starknet):
     return await starknet.deploy(
-        source="./tests/unit/kakarot/test_execution_context.cairo",
+        source="./tests/unit/src/kakarot/test_execution_context.cairo",
         cairo_path=["src"],
         disable_hint_validation=True,
     )
