+++ tests/unit/src/kakarot/test_stack.cairo
