+++ tests/unit/src/kakarot/test_stack.py
@@ -7,7 +7,7 @@ import pytest_asyncio
 @pytest_asyncio.fixture
 async def stack(starknet):
     return await starknet.deploy(
-        source="./tests/unit/kakarot/test_stack.cairo",
+        source="./tests/unit/src/kakarot/test_stack.cairo",
         cairo_path=["src"],
         disable_hint_validation=True,
     )
